requirements.txt
appnope==0.1.0
backcall==0.1.0
beautifulsoup4==4.6.3
bleach==3.0.2
boto3==1.9.48
botocore==1.12.48
bs4==0.0.1
cachetools==3.0.0
certifi==2018.10.15
chardet==3.0.4
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
Django==2.1.3
docutils==0.14
entrypoints==0.2.3
et-xmlfile==1.0.1
gevent==1.3.7
google-api-core==1.5.2
google-api-python-client==1.7.4
google-auth==1.6.1
google-auth-httplib2==0.0.3
google-cloud==0.34.0
google-cloud-vision==0.35.1
googleapis-common-protos==1.5.5
greenlet==0.4.15
grpcio==1.16.1
httplib2==0.12.0
idna==2.7
image==1.5.27
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
jdcal==1.4
jedi==0.13.1
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==3.0.1
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
nltk==3.3
notebook==5.7.0
numpy==1.15.3
openpyxl==2.5.9
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prometheus-client==0.4.2
prompt-toolkit==2.0.7
protobuf==3.6.1
ptyprocess==0.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycosat==0.6.3
Pygments==2.2.0
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
pyzmq==17.1.2
qtconsole==4.4.2
requests==2.20.0
rsa==4.0
ruamel.yaml==0.15.76
s3transfer==0.1.13
scikit-learn==0.20.0
scipy==1.1.0
seaborn==0.9.0
selenium==3.141.0
Send2Trash==1.5.0
six==1.11.0
sklearn==0.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
tqdm==4.28.1
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.24
virtualenv==16.1.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
xlrd==1.1.0
